+++ app.ts
@@ -3,6 +3,7 @@ import bodyParser from "body-parser";
 
 import adminRoutes from "./routes/admin.route";
 import shopRoutes from "./routes/shop.route";
+import cartRoutes from "./routes/cart.route";
 
 const app = express();
 
@@ -11,6 +12,7 @@ app.use(bodyParser.urlencoded({ extended: false }));
 
 app.use("/admin", adminRoutes);
 app.use("/shop", shopRoutes);
+app.use("/cart", cartRoutes);
 
 app.use((_req, res) => {
   res.status(404).json({